backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-storages[boto3]
dj-database-url
gunicorn
